moved tests to appropriate scripts
