+++ tests/test_report.py
@@ -11,9 +11,6 @@ from sklearn.model_selection import KFold
 import pandas as pd
 import numpy as np
 
-from xgboost import XGBClassifier
-from lightgbm import LGBMClassifier
-
 import string
 import random
 
@@ -38,131 +35,6 @@ def create_data(classification=True):
 
 df = create_data()
 df_r = create_data(classification=False)
-
-
-def test_get_coef():
-    y = df['target']
-    df_1 = df.drop('target', axis=1)
-    
-    full_pipe = Pipeline([('scaler', tml.DfScaler()), 
-                          ('logit', LogisticRegression(solver='lbfgs', multi_class='auto'))])
-    
-    full_pipe.fit(df_1, y)
-
-    with pytest.warns(None) as record:
-        coef = tml.get_coef(full_pipe)
-    assert len(record) == 0
-    assert len(coef) == df_1.shape[1]
-
-
-@pytest.mark.parametrize('model', [DecisionTreeClassifier(), 
-                                   XGBClassifier(use_label_encoder=False), 
-                                   LGBMClassifier()])    
-def test_feat_imp(model):
-    '''
-    Test if we can get the feature importance for various models
-    XGB - Version 1.3.3, not in package requirements
-    LGB - Version 3.1.1, not in package requirements
-    '''
-    y = df['target']
-    df_1 = df.drop('target', axis=1)
-    
-    full_pipe = Pipeline([('scaler', tml.DfScaler()), 
-                          ('model', model)])
-    
-    full_pipe.fit(df_1, y)
-
-    with pytest.warns(None) as record:
-        coef = tml.get_feature_importance(full_pipe)
-    assert len(record) == 0
-    assert len(coef) == df_1.shape[1]
-
-
-@patch("matplotlib.pyplot.show")
-def test_learning_curves(_):
-    '''
-    Test learning curves can be plotted
-    '''
-    y = df['target']
-    df_1 = df.drop('target', axis=1)
-    
-    full_pipe = Pipeline([('scaler', tml.DfScaler()), 
-                          ('logit', LogisticRegression(solver='lbfgs', multi_class='auto'))])
-    
-    kfold = KFold(n_splits=3)
-    with pytest.warns(None) as record:
-        tml.plot_learning_curve(estimator=full_pipe, X=df_1, y=y, scoring='accuracy', ylim=(0, 1), cv=kfold,
-                            n_jobs=-1, train_sizes=np.linspace(.1, 1.0, 10), title='Title')
-    assert len(record) == 0
-    
-    
-@patch("matplotlib.pyplot.show")  
-def test_learning_curves_xgb(mock_show):
-    '''
-    Test learning curves can be plotted with xbgboost
-    '''
-    y = df['target']
-    df_1 = df.drop('target', axis=1)
-    
-    full_pipe = Pipeline([('scaler', tml.DfScaler()), 
-                          ('xgb', XGBClassifier(objective='binary:logistic', 
-                                                n_estimators=2, n_jobs=-1,
-                                                use_label_encoder=False))])
-    
-    kfold = KFold(n_splits=3)
-    with pytest.warns(None) as record:
-        tml.plot_learning_curve(estimator=full_pipe, X=df_1, y=y, scoring='accuracy', ylim=(0, 1), cv=kfold,
-                            n_jobs=-1, train_sizes=np.linspace(.1, 1.0, 10), title=None)
-    assert len(record) == 0   
-    
-    
-@patch("matplotlib.pyplot.show")
-def test_learning_curves_lgb(_):
-    '''
-    Test learning curves can be plotted with xbgboost
-    '''
-    y = df['target']
-    df_1 = df.drop('target', axis=1)
-    
-    full_pipe = Pipeline([('scaler', tml.DfScaler()), 
-                          ('lgb', LGBMClassifier())])
-    
-    kfold = KFold(n_splits=3)
-    with pytest.warns(None) as record:
-        tml.plot_learning_curve(estimator=full_pipe, X=df_1, y=y, scoring='accuracy', ylim=None, cv=kfold,
-                            n_jobs=-1, train_sizes=np.linspace(.1, 1.0, 10), title=None)
-    assert len(record) == 0
-
-
-@patch("matplotlib.pyplot.show")
-def test_plot_feat_imp(_):
-    '''
-    Test if plot feat importance works
-    '''
-    y = df['target']
-    df_1 = df.drop('target', axis=1)
-    
-    full_pipe = Pipeline([('scaler', tml.DfScaler()), 
-                          ('lgb', LGBMClassifier())])
-    
-    kfold = KFold(n_splits=3)
-    full_pipe.fit(df_1, y)
-    oof, coef = tml.cv_score(df_1, y, full_pipe, kfold, imp_coef=True, predict_proba=False)
-
-    with pytest.warns(None) as record:
-        tml.plot_feat_imp(coef['feat_imp'])
-    assert len(record) == 0
-    
-
-def test_plot_feat_imp_warning():
-    '''
-    Test if plot feat importance works
-    '''
-    wrong_input = pd.DataFrame({'a': [1,2,3], 
-                                'b': [2,3,4]})
-
-    with pytest.raises(KeyError):
-        tml.plot_feat_imp(wrong_input, n=10)
 
 
 @patch("matplotlib.pyplot.show")
